refactor nytimes to extract xml utilities
