+++ feed/sources/nytimes.py
@@ -1,55 +1,19 @@
-from io import StringIO
+from typing import Dict, List
 from xml.etree import ElementTree
 from feed.articles.NewsArticleImage import NewsArticleImage
 from feed.articles.NewsArticleSummary import NewsArticleSummary
 from feed.utils.request import get_xml
-
-def extract_text_from_tag(xml: ElementTree.Element, tag_name: str, name_space=None):
-    """
-    Extracts the text of the given tag. If the tag does not exist it returns empty string.
-
-    Args:
-        xml: Element that contains the tag
-        tag_name: name of the tag to extract text from
-
-    Returns:
-        Text with in the tag
-    """
-    value = xml.find(tag_name, name_space)
-    if value is None:
-        return ''
-    return value.text
-
-def extract_text_from_attr(xml: ElementTree.Element, attr_name: str):
-    """
-    Extracts the text of the given tag. If the tag does not exist it returns empty string.
-
-    Args:
-        xml: Element that contains the tag
-        tag_name: name of the tag to extract text from
-
-    Returns:
-        Text with in the tag
-    """
-    try:
-        value = xml.attrib.get('url')
-        if value is None:
-           return ''
-        return value
-    except AttributeError:
-        return ''
-    
-
-# def get_namespaces(schema):
-#     return dict([ node for _, node in schema])
+from feed.utils.xml import (
+    get_namespaces,
+    extract_text_from_attr,
+    extract_text_from_tag
+)
 
 def mapper(item: ElementTree.Element, name_spaces) -> NewsArticleSummary:
     """
         Converts xml Element of item to NewsArticleSummary
     """
     summary = NewsArticleSummary()
-    # print(item.tag)
-    # name_spaces = get_namespaces(item)
     summary.title = extract_text_from_tag(item, 'title')
     summary.link = extract_text_from_tag(item, 'link')
     summary.published_on = extract_text_from_tag(item, 'pubDate')
@@ -62,33 +26,22 @@ def mapper(item: ElementTree.Element, name_spaces) -> NewsArticleSummary:
     summary.image = NewsArticleImage(image_url, image_credit, image_alt)
     return summary
 
-def parse(xml_string, name_spaces):
-    """Given the full """
+def parse(xml_string: ElementTree.Element, name_spaces: Dict[str, str]) -> List[NewsArticleSummary]:
+    """Given the full XML of NY times feed, it returns list of NewsArticleSummary"""
     channel = xml_string.findall('channel')[0]
     return list(map(lambda x: mapper(x, name_spaces), channel.findall('item')))
         
-def get_namespaces(xml_string):
-    """
-        Given the xml string finds all the namespaces for the xml.
-        Example: 
-            Given:
-                <rss xln:dc="abc" xln:media="xyz">....</rss>
-            Returns:
-                {'dc': 'abc', 'media': 'zyz'}
-    """
-    namespaces = dict([
-            node for _, node in ElementTree.iterparse(
-                StringIO(xml_string), events=['start-ns']
-            )
-    ])
-    return namespaces
 
-def get_today():
+
+def get_home() -> List[NewsArticleSummary]:
+    """
+    Fetches list of news articles published on NYTime's HomePage
+    """
     ny_times_home, schema = get_xml('https://rss.nytimes.com/services/xml/rss/nyt/HomePage.xml')
     name_spaces = get_namespaces(schema)
     return parse(ny_times_home.getroot(), name_spaces)
 
 # Using a CMD tool to see today's news
 if __name__ == "__main__":
-    articles = get_today()
+    articles = get_home()
     print(articles)
