+++ setup.py
@@ -11,7 +11,7 @@ print(dependencies)
 
 setup(
     name='feed',
-    version="0.0.3",
+    version="0.0.4",
     long_description=long_description,
     long_description_content_type="text/markdown",
     author="Upen Dhakal<dhakal.upenn@gmail.com>",
